bug fixes: 1. in func check_retrace_when_short(), -    tick_4_high = rates[3]['high'] +    tick_4_high = rates[4]['high'] tick_4_high was wrong assigned rates[3]['high'], when it should be rates[4]['high']
2. in funcs check_steps_when_long/short(), we should compare the nearer high/low,
in the past we returned the farther wrong high/low

Note: these funcs are still NOT functioning as we expected. so I DISABLED using them as filters

3. the offset limit only prevents opening an order immediately
when tp is met on the candle where tp is taken, but it cannot prevent opening
another order on the next candle passing its previous 2 candles

temporary workaround:
we now use a timer to count down for 2 ticks after the sl is moved to breakeven
this in most situations prevents wrongly opening orders

4. added checking tick_two_close > tick_two_open to ensure this key candle crossing sma is closed a bullish candle.
in below to above sma across, in the crossing candle there might be a jump (window) or maybe the sma is steep, and the candle will close as a bearish candle with a very small body and a long upper wick, but its main body is above sma, and it closes above sma
if this happens, we do not think this is a valid crossing. i guess this usually happens in lower timeframes. i observed this on m5

applied this also for bearish crossing

changes:
1. moved constants including account_live, account_demo, and path to the main() function.

2. use the new calculate_lot_size function.
  a. this function calculates the conservative lot size
  b. calculates the exactly pip value for any currency pair
  c. check if the lot size is < 0.01, if so then quit mt5 and quit the program
  b. if lot size > 100, then change it to 100

3. if buy, use the ask price mt5.symbol_info_tick(symbol).ask. if sell, use the bid price mt5.symbol_info_tick(symbol).bid

4. changed magic number and comment

5. added a function check_if_its_trading_time()

6. moved below to above, so that they do not get calculated multiple times
digits = mt5.symbol_info(symbol).digits
multiply_digits = 10 ** digits

7. disabled closing order after counting down

8. disabled timer

9. added func select_account()